+++ Actual_Website/ERP_Agnus_Dei_School_Systems_INC/app/Mail/InquiryVerificationMail.php
@@ -10,7 +10,7 @@ use Illuminate\Mail\Mailables\Content;
 use Illuminate\Mail\Mailables\Envelope;
 use Illuminate\Queue\SerializesModels;
 
-class InquiryVerificationMail extends Mailable
+class InquiryVerificationMail extends Mailable implements ShouldQueue
 {
     use Queueable, SerializesModels;